lights: update timer trigger control to fix LED blinking issue
Update control logic to set "timer" trigger and corresponding
"delay_on" and "delay_off" parameters for LED blinking function.

Following changes has been made:

1.clear timer trigger before setting brightness:
If set LED brightness to a none-zero value when it's
enabled with "timer" trigger, the "timer" trigger won't
be cleared automatically and the LED would be still in
working blink mode. To avoid this, clear "timer" trigger
before setting brightness values.

2.set timer trigger before updating delay_on/off values:
For a LED class device that uses timer trigger and has
default "delay_on" and "delay_off" values, the LED device
will be enabled in blink mode immediately once the "timer"
trigger is updated. So write the "timer" trigger as the
1st step before enabling the blink mode and following
with the "delay_on" "delay_off" writes. Meanwhile, "delay_on"
and "delay_off" files are created after writing "timer" to
"trigger" in a LED class device runtime. Ueventd needs
to update the system permission for this LED class device
which would take some time. Hence writing immediately to
delay_{on/off} would end up in access denial. To avoid
this, add a retry logic with delay so that they can be
written successfully.

Change-Id: Ied2396b68141d3b1d5acd67d310dadac8586e840