lights: Update QMAA logic
Move QMAA logic below the light ID check so that the calling with any
invalid light ID would be caught and bail out with UNSUPPORTED_OPERATION
error code. This ensures the light HAL VTS test passes in the case where
an invalid ID is passed when there is no real light HW present.

Also update the QTI copyright to the latest format by removing
the year and author email.

Change-Id: I82fe4af735dae82fead7311bcb293949933592f6